Create folder for components and separate folder for css. Update imports
