+++ src/CSS/Nutrition.css
@@ -1,0 +1,4 @@
+.nutrition-container {
+  display: flex;
+  
+}
